fix: address Warden review issues in salary package
- Add store_test.go with unit tests for all store functions (GetConfig,
  SaveConfig, SeedDefaultTiers, GetCommissionTiers, SaveRecord,
  GetRecords, GetTaxBrackets, GetHoursWorked)
- Change INTEGER struct fields from int to int64 (TaxBracket.Year,
  Record.WorkingDays/VacationDays/SickDays) to match SQLite column types
- Remove redundant indexes idx_salary_config_user_id and
  idx_salary_records_user_id (covered by existing UNIQUE constraints)
- GetHoursWorked now returns errors on time parse failure instead of
  silently skipping malformed rows

Bead: Hytte-6y0o

+++ internal/salary/calculator.go
@@ -86,7 +86,7 @@ func EstimateMonth(
 
 	return Record{
 		UserID:        config.UserID,
-		WorkingDays:   workingDays,
+		WorkingDays:   int64(workingDays),
 		HoursWorked:   hoursWorked,
 		BillableHours: hoursWorked,
 		BaseAmount:    baseAmount,
@@ -94,8 +94,8 @@ func EstimateMonth(
 		Gross:         gross,
 		Tax:           tax,
 		Net:           net,
-		VacationDays:  vacationDays,
-		SickDays:      sickDays,
+		VacationDays:  int64(vacationDays),
+		SickDays:      int64(sickDays),
 		IsEstimate:    true,
 	}
 }
